# SWAGGER_OPENAPI.yaml
openapi: 3.0.0
info:
  title: Sprotif Adherent Service API
  version: 1.0.0
  description: Complete API documentation for the Sprotif adherent and authentication microservice
  contact:
    name: Sprotif Team
    email: support@sprotif.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8081
    description: Local Development Server
  - url: http://localhost:3000
    description: API Gateway

tags:
  - name: Authentication
    description: User authentication and registration endpoints
  - name: Profile
    description: User profile and subscription management

paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid registration data
        '409':
          description: User already exists

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful, JWT token returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
        '400':
          description: Bad request

  /api/auth/admin/register:
    post:
      tags:
        - Authentication
      summary: Register a new admin
      description: Create a new admin account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Admin registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid registration data
        '409':
          description: User already exists

  /api/profile:
    get:
      tags:
        - Profile
      summary: Get current user profile
      description: Retrieve the profile and subscription information of the logged-in user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Unauthorized - missing or invalid JWT token
        '403':
          description: Forbidden - user does not have access

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: SecurePassword123!
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: SecurePassword123!

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT bearer token for authentication
          example: eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJ...
        username:
          type: string
          example: user@example.com
        email:
          type: string
          format: email
          example: user@example.com
        role:
          type: string
          enum: [USER, ADMIN]
          example: USER

    Adherent:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        phoneNumber:
          type: string
          example: "+212698273901"
        dateOfBirth:
          type: string
          format: date
          example: 1990-01-15
        address:
          type: string
          example: 123 Main St
        city:
          type: string
          example: Paris
        postalCode:
          type: string
          example: "75001"
        country:
          type: string
          example: France
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, EXPIRED, DEACTIVATED]
          example: ACTIVE
        subscriptionType:
          type: string
          enum: [BASIC, PREMIUM]
          nullable: true
          example: PREMIUM
        subscriptionPrice:
          type: number
          format: double
          nullable: true
          example: 49.95
        subscriptionStartDate:
          type: string
          format: date
          nullable: true
          example: 2026-02-19
        subscriptionDurationMonths:
          type: integer
          nullable: true
          example: 12
        subscriptionEndDate:
          type: string
          format: date
          nullable: true
          example: 2027-02-19

    Subscription:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        adherentId:
          type: integer
          format: int64
          example: 1
        type:
          type: string
          enum: [BASIC, PREMIUM]
          example: PREMIUM
        price:
          type: number
          format: double
          example: 49.95
        weeklySessions:
          type: integer
          example: 10
        weeklySessionsUsed:
          type: integer
          example: 2
        durationMonths:
          type: integer
          example: 12
        startDate:
          type: string
          format: date-time
          example: 2026-02-19T00:00:00Z
        endDate:
          type: string
          format: date-time
          example: 2027-02-19T00:00:00Z
        active:
          type: boolean
          example: true

    ProfileResponse:
      type: object
      properties:
        adherent:
          $ref: '#/components/schemas/Adherent'
        subscription:
          $ref: '#/components/schemas/Subscription'
          nullable: true